+++ src/components/map/navigraph/NavigraphAirways.vue
@@ -9,6 +9,10 @@ import { Point } from 'ol/geom';
 import type { ShallowRef } from 'vue';
 import type VectorSource from 'ol/source/Vector';
 import greatCircle from '@turf/great-circle';
+import { useMapStore } from '~/store/map';
+import type { ObjectWithGeometry } from 'ol/Feature';
+import type { Coordinate } from 'ol/coordinate';
+import { checkFlightLevel } from '~/composables/data';
 
 defineSlots<{ default: () => any }>();
 
@@ -16,19 +20,29 @@ const source = inject<ShallowRef<VectorSource>>('navigraph-source');
 
 const store = useStore();
 const dataStore = useDataStore();
+const mapStore = useMapStore();
 
 const isEnabled = computed(() => store.mapSettings.navigraphData?.airways?.enabled !== false);
+let geometries: ObjectWithGeometry[] = [];
+let waypointGeometries: ObjectWithGeometry[] = [];
 let features: Feature[] = [];
 
-watch(isEnabled, async val => {
-    if (!val) {
+const extent = computed(() => mapStore.extent);
+const level = computed(() => store.mapSettings.navigraphData?.mode);
+
+watch([isEnabled, extent, level], async ([enabled, extent]) => {
+    if (!enabled) {
         source?.value.removeFeatures(features);
         features = [];
+        geometries = [];
+        waypointGeometries = [];
+        return;
     }
-    else {
-        const entries = Object.entries(dataStore.navigraph.data.value!.airways);
-        const len = entries.length;
 
+    const entries = Object.entries(dataStore.navigraph.data.value!.airways);
+    const len = entries.length;
+
+    if (!geometries.length) {
         for (let i = 0; i < len; i += 100) {
             for (let k = i; k < i + 100; k++) {
                 const entry = entries[k];
@@ -36,21 +50,24 @@ watch(isEnabled, async val => {
                 const [key, [identifier, type, waypoints]] = entry;
 
                 if (type === 'C') continue;
+
                 waypoints.forEach((waypoint, index) => {
                     const nextWaypoint = waypoints[index + 1];
 
-                    features.push(new Feature({
-                        geometry: new Point([waypoint[3], waypoint[4]]),
+                    waypointGeometries.push({
+                        rawCoords: [waypoint[3], waypoint[4]],
                         key,
                         identifier,
+                        flightLevel: waypoint[5],
                         waypoint: waypoint[0],
                         type: 'airway-waypoint',
-                    }));
+                        dataType: 'navdata',
+                    });
 
                     if (!nextWaypoint) return;
 
-                    features.push(new Feature({
-                        geometry: turfGeometryToOl(greatCircle([waypoint[3], waypoint[4]], [nextWaypoint[3], nextWaypoint[4]], { npoints: 3 })),
+                    geometries.push({
+                        rawCoords: [[waypoint[3], waypoint[4]], [nextWaypoint[3], nextWaypoint[4]]],
                         key,
                         identifier,
                         inbound: waypoint[1],
@@ -58,18 +75,37 @@ watch(isEnabled, async val => {
                         waypoint: waypoint[0],
                         flightLevel: waypoint[5],
                         type: 'airways',
-                    }));
+                        dataType: 'navdata',
+                    });
                 });
             }
 
             await sleep(0);
         }
-
-        source?.value.addFeatures(features);
     }
+
+    const newFeatures = [
+        ...geometries.filter(x => checkFlightLevel(x.flightLevel) && x.rawCoords.some((x: Coordinate) => isPointInExtent(x, extent))).map(x => new Feature({
+            ...x,
+            geometry: turfGeometryToOl(greatCircle(x.rawCoords[0], x.rawCoords[1], { npoints: 3 })),
+        })),
+        ...waypointGeometries.filter(x => checkFlightLevel(x.flightLevel) && isPointInExtent(x.rawCoords, extent)).map(x => new Feature({
+            ...x,
+            geometry: new Point(x.rawCoords),
+        })),
+    ];
+
+    source?.value.removeFeatures(features);
+    features = newFeatures;
+    source?.value.addFeatures(features);
 }, {
     immediate: true,
 });
 
-onBeforeUnmount(() => source?.value.removeFeatures(features));
+onBeforeUnmount(() => {
+    source?.value.removeFeatures(features);
+    geometries = [];
+    waypointGeometries = [];
+    features = [];
+});
 </script>