+++ src/components/map/navigraph/NavigraphNdb.vue
@@ -8,12 +8,14 @@ import { Feature } from 'ol';
 import { Point } from 'ol/geom';
 import type { ShallowRef } from 'vue';
 import type VectorSource from 'ol/source/Vector';
+import { useMapStore } from '~/store/map';
 
 defineSlots<{ default: () => any }>();
 
 const source = inject<ShallowRef<VectorSource>>('navigraph-source');
 
 const store = useStore();
+const mapStore = useMapStore();
 const dataStore = useDataStore();
 
 const isNDBEnabled = computed(() => store.mapSettings.navigraphData?.ndb !== false);
@@ -21,34 +23,36 @@ const isVorEnabled = computed(() => store.mapSettings.navigraphData?.vordme !== 
 let ndb: Feature[] = [];
 let vordme: Feature[] = [];
 
-watch([isNDBEnabled, isVorEnabled], () => {
-    if (!isNDBEnabled.value) {
-        source?.value.removeFeatures(ndb);
-        ndb = [];
-    }
-    else if (!ndb.length) {
-        ndb = Object.entries(dataStore.navigraph.data.value!.ndb).map(([key, [name, code, frequency, longitude, latitude]]) => new Feature({
+const extent = computed(() => mapStore.extent);
+
+watch([isNDBEnabled, isVorEnabled, extent, dataStore.navigraph.data], () => {
+    source?.value.removeFeatures(ndb);
+    ndb = [];
+
+    source?.value.removeFeatures(vordme);
+    vordme = [];
+
+    if (isNDBEnabled.value && dataStore.navigraph.data.value?.ndb) {
+        ndb = Object.entries(dataStore.navigraph.data.value.ndb).filter(([, x]) => isPointInExtent([x[3], x[4]], extent.value)).map(([key, [name, code, frequency, longitude, latitude]]) => new Feature({
             geometry: new Point([longitude, latitude]),
             key,
             name,
             code,
             frequency,
+            dataType: 'navdata',
             type: 'ndb',
         }));
         source?.value.addFeatures(ndb);
     }
 
-    if (!isVorEnabled.value) {
-        source?.value.removeFeatures(vordme);
-        vordme = [];
-    }
-    else if (!vordme.length) {
-        vordme = Object.entries(dataStore.navigraph.data.value!.vhf).map(([key, [name, code, frequency, longitude, latitude]]) => new Feature({
+    if (isVorEnabled.value && dataStore.navigraph.data.value?.vhf) {
+        vordme = Object.entries(dataStore.navigraph.data.value.vhf).filter(([, x]) => isPointInExtent([x[3], x[4]], extent.value)).map(([key, [name, code, frequency, longitude, latitude]]) => new Feature({
             geometry: new Point([longitude, latitude]),
             key,
             name,
             code,
             frequency,
+            dataType: 'navdata',
             type: 'vhf',
         }));
         source?.value.addFeatures(vordme);