+++ src/server/api/data/navigraph/procedure/[type]/[airport]/[procedure]/short.ts
@@ -1,0 +1,5 @@
+import { getNavDataProcedure } from '~/utils/backend/navigraph/navdata';
+
+export default defineEventHandler(async event => {
+    return getNavDataProcedure(event, 'short');
+});